% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{area}
\alias{area}
\title{Area of a circle}
\usage{
area(shape)
}
\arguments{
\item{shape}{The shape as a BaseCircle class object.}
}
\value{
\code{NULL}
}
\description{
This is a generic function.
\code{area} calculates and prints the area of a circle.
}
