% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_circle.R
\docType{class}
\name{BaseCircle-class}
\alias{BaseCircle-class}
\title{A Base Circle object.}
\description{
A Base Circle object.
}
\section{Slots}{

\describe{
\item{\code{r}}{The radius as a number.}

\item{\code{d}}{The diameter as a number.}

\item{\code{name}}{The name of the object as a character string.}
}}

