% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_circle.R
\docType{class}
\name{ColorCircle-class}
\alias{ColorCircle-class}
\title{A Color Circle object that inherits from a BaseCircle object.}
\description{
A Color Circle object that inherits from a BaseCircle object.
}
\section{Slots}{

\describe{
\item{\code{color}}{The color of the object as a character string.}
}}

