% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{perimeter}
\alias{perimeter}
\title{Perimeter of a circle}
\usage{
perimeter(shape)
}
\arguments{
\item{shape}{The shape as a BaseCircle class object.}
}
\value{
\code{NULL}
}
\description{
This is a generic function.
\code{perimeter} calculates and prints the perimeter of a circle.
}
