% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_diameter}
\alias{set_diameter}
\title{Sets the diameter of the circle.}
\usage{
set_diameter(shape)
}
\arguments{
\item{shape}{The shape as a BaseCircle class object.}
}
\description{
This is a generic function.
\code{set_diameter} sets the diameter of the BaseCircle object
by populating the \code{d} slot.
}
